Added ability to handle tests & suite setup that return promises.
With this change, if a suite or test function returns a promise,
joe will then wait for that promise to complete before continuing
without requiring that you take a completion callback or call it.

This simplifies testing with promise-based APIs and also lets you
use async/await inside test function and suite setup.